More tests completed

+++ src/main/java/rocks/zipcode/assessment2/collections/MonthConversion.java
@@ -36,7 +36,7 @@ public class MonthConversion {
      * @param monthName - name of month
      * @return - the ordinal of the month in the year
      */
-    public int getNumber(String monthName) {
+    public Integer getNumber(String monthName) {
         for(Object o: calendarMap.keySet()){
             if(calendarMap.get(o).equals(monthName)){
                 return (Integer) o;
@@ -46,7 +46,7 @@ public class MonthConversion {
 
         }
 
-       return (Integer) null;
+       return null;
     }
 
     /**
